requirements.txt
dgl-cu101==0.6.1
ogb==1.3.1
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1